# GS: Use 16bit RGBA unorm for HDR