--- a/bin/resources/shaders/opengl/convert.glsl
+++ b/bin/resources/shaders/opengl/convert.glsl
@@ -322,7 +322,7 @@ void ps_datm0()
 void ps_hdr_init()
 {
     vec4 value = sample_c();
-    SV_Target0 = vec4(round(value.rgb * 255.0f), value.a);
+    SV_Target0 = vec4(round(value.rgb * 255.0f) / 65535.0f, value.a);
 }
 #endif
 
@@ -330,7 +330,7 @@ void ps_hdr_init()
 void ps_hdr_resolve()
 {
     vec4 value = sample_c();
-    SV_Target0 = vec4(vec3(ivec3(value.rgb) & 255) / 255.0f, value.a);
+    SV_Target0 = vec4(vec3(uvec3(value.rgb * 65535.0f) & 255u) / 255.0f, value.a);
 }
 #endif