--- a/pcsx2/GS/Renderers/Common/GSTexture.cpp
+++ b/pcsx2/GS/Renderers/Common/GSTexture.cpp
@@ -84,7 +84,7 @@ u32 GSTexture::GetCompressedBytesPerBlock() const
 	static constexpr u32 bytes_per_block[] = {
 		1, // Invalid
 		4, // Color/RGBA8
-		16, // HDRColor/RGBA32F
+		8, // HDRColor/RGBA16
 		32, // DepthStencil
 		1, // UNorm8/R8
 		2, // UInt16/R16UI