--- a/pcsx2/GS/Renderers/Metal/tfx.metal
+++ b/pcsx2/GS/Renderers/Metal/tfx.metal
@@ -784,17 +784,6 @@ struct PSMain
 
 	void ps_color_clamp_wrap(thread float4& C)
 	{
-		if (PS_HDR && PS_COLCLIP) // COLCLIP flag indicates accumulation blend under HDR
-		{
-			int3 color = int3(C.rgb);
-			if (PS_DFMT == FMT_16)
-				color &= 0xF8;
-			// -128 to 127 gives us longer before we run out of float precision
-			// Especially for games that mainly use 1 and 255 (sly), since that maps to 1 and -1
-			C.rgb = float3(char3(color));
-			return;
-		}
-
 		// When dithering the bottom 3 bits become meaningless and cause lines in the picture so we need to limit the color depth on dithered items
 		if (!SW_BLEND && !PS_DITHER && !PS_FBMASK)
 			return;
@@ -812,7 +801,7 @@ struct PSMain
 		if (PS_DFMT == FMT_16 && PS_BLEND_MIX == 0)
 			// In 16 bits format, only 5 bits of colors are used. It impacts shadows computation of Castlevania
 			C.rgb = float3(short3(C.rgb) & 0xF8);
-		else if (PS_COLCLIP && !PS_HDR)
+		else if (PS_COLCLIP || PS_HDR)
 			C.rgb = float3(short3(C.rgb) & 0xFF);
 	}
 
@@ -989,7 +978,7 @@ struct PSMain
 		ps_fbmask(C);
 
 		if (PS_COLOR0)
-			out.c0 = PS_HDR ? float4(C.rgb, C.a / 255.f) : C / 255.f;
+			out.c0 = PS_HDR ? float4(C.rgb / 65535.f, C.a / 255.f) : C / 255.f;
 		if (PS_COLOR0 && PS_ONLY_ALPHA)
 			out.c0.rgb = 0;
 		if (PS_COLOR1)