--- a/pcsx2/GS/Renderers/OpenGL/GSTextureOGL.cpp
+++ b/pcsx2/GS/Renderers/OpenGL/GSTextureOGL.cpp
@@ -225,10 +225,10 @@ GSTextureOGL::GSTextureOGL(Type type, int width, int height, int levels, Format 
 
 		// 4 channel float
 		case Format::HDRColor:
-			gl_fmt          = GL_RGBA32F;
+			gl_fmt          = GL_RGBA16;
 			m_int_format    = GL_RGBA;
-			m_int_type      = GL_FLOAT;
-			m_int_shift     = 4;
+			m_int_type      = GL_UNSIGNED_SHORT;
+			m_int_shift     = 3;
 			break;
 
 		// Depth buffer