### Description of Changes
It looks like #7131 wasn't enough to make half floats viable
So we're trying with 16-bit unorm instead
Should allow up to 256 overlaps all laying down 255s, or some slightly larger number of smaller values

Alternative to #7132
May break games that use HDR without accumulation blends, we haven't found any yet, but there's always going to be that one game...

### Rationale behind Changes
Improved performance in HDR

### Suggested Testing Steps
Test HDR games
Test OGL and DX (I only tested Metal and Vk)