@lightningterror 2022-10-09
[BigMuthaTruckers.zip](https://github.com/PCSX2/pcsx2/files/9741041/BigMuthaTruckers.zip)
[Castlevania shadow.zip](https://github.com/PCSX2/pcsx2/files/9741043/Castlevania.shadow.zip)
[Lord of the rings Third Age.zip](https://github.com/PCSX2/pcsx2/files/9741044/Lord.of.the.rings.Third.Age.zip)

Here are some dumps where hdr is enabled without accumulation blend.
For Castlevania and Big Mutha you will need to force disable barriers.