@refractionpcsx2 2022-10-09
Here's dumps which are usually affected by HDR, all seem to work absolutely fine on this PR! (booo to floats! down with floats!)

[narnia.zip](https://github.com/PCSX2/pcsx2/files/9741563/narnia.zip)
[sly2.zip](https://github.com/PCSX2/pcsx2/files/9741565/sly2.zip)
[evolutionsnowboarding.zip](https://github.com/PCSX2/pcsx2/files/9741567/evolutionsnowboarding.zip)
[Full spectrum warrior.zip](https://github.com/PCSX2/pcsx2/files/9741568/Full.spectrum.warrior.zip)