@refractionpcsx2 2022-10-09
I went through ALL my dumps, no regressions, and all the speed improvements of #7132 
Good job :)