@tellowkrinkle 2022-10-09
Big Mutha Truckers uses 1202 (Cd * As) which should run fine as long as alpha is ≤1.0.  The trace uses it for darkening, but that doesn't mean the game doesn't use it for brightening elsewhere.
Castlevania Shadow uses 0111 (Normal alpha blend equation but with Ad), which will be equally broken on both float and unorm
LotR uses 2101 (Cd - As * Cd) which should run fine as long as alpha is ≤1.0.  The trace uses 1.0 and 0.25 alpha with an RGB24 texture and AEM, so we should be fine here.