@GN876 2022-10-11
on an RTX 3080, I got 2x the framerate in Sly 2 at 6x native. Framerates in Echo Night: Beyond are also better (around 30% I think) compared to latest master. In both games I didn't see even a single visual issue.