@lightningterror 2022-10-11
Here are some benchmarks with my 750ti on 3x native:

Master:
```
Full Spectrum Warrior
OpenGL 93 fps
Vulkan 98 fps
Direct3D11 95 fps
Direct3D12 95 fps

Sly 2
OpenGL 41 fps
Vulkan 46 fps
Direct3D11 39 fps
Direct3D12 40 fps
```

PR:
```
Full Spectrum Warrior
OpenGL 134 fps
Vulkan 141 fps
Direct3D11 134 fps
Direct3D12 125 fps

Sly 2
OpenGL 74 fps
Vulkan 92 fps
Direct3D11 76 fps
Direct3D12 78 fps
```